Make grabRogueWindows relative to their workspace
This patch suggests that, for example, if num_workspaces is 5 and we
have 2 monitors, and assuming we have a window on workspace 7 (the
second workspace on the second monitor), then when the second monitor
is removed, the window will be grabbed to workspace 2 (and not the
currently active monitor).

This makes the grab notion behave more like Microsoft Windows ( :-( )
but is more friendly to users.